remove extra (previous version) handler file, rename newer handler
